requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
